refactor(library): Change all `*.h` files to `*.hpp`.
This should help associate every header file to C++.
