+++ src/lib/CRSFforArduino/src/CRSFforArduino.hpp
@@ -1,7 +1,7 @@
 /**
- * @file CRSF.h
+ * @file CRSFforArduino.hpp
  * @author Cassandra "ZZ Cat" Robinson (nicad.heli.flier@gmail.com)
- * @brief CRSF class definition.
+ * @brief CRSF for Arduino facilitates the use of ExpressLRS RC receivers in Arduino projects.
  * @version 0.5.0
  * @date 2023-10-24
  *
@@ -21,49 +21,40 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with CRSF for Arduino.  If not, see <https://www.gnu.org/licenses/>.
- * 
+ *
  */
 
 #pragma once
 
 #include "Arduino.h"
-#include "CRSFProtocol.h"
-#if defined(ARDUINO) && defined(PLATFORMIO)
-#ifdef USE_DMA
-#include "Hardware/DevBoards/DevBoards.h"
-#endif
-#include "SerialReceiver/CRC/CRC.h"
-#elif defined(ARDUINO) && !defined(PLATFORMIO)
-#ifdef USE_DMA
-#include "lib/CRSFforArduino/src/Hardware/DevBoards/DevBoards.h"
-#endif
-#include "lib/CRSFforArduino/src/SerialReceiver/CRC/CRC.h"
-#endif
-// #include "Hardware.h"
 
-namespace serialReceiver
-{
-    class CRSF
-#ifdef USE_DMA
-        : private hal::DevBoards
+#if defined(ARDUINO) && defined(PLATFORMIO)
+#include "SerialReceiver/SerialReceiver.hpp"
+#elif defined(ARDUINO) && !defined(PLATFORMIO)
+#include "lib/CRSFforArduino/src/SerialReceiver/SerialReceiver.hpp"
 #endif
+
+namespace sketchLayer
+{
+    class CRSFforArduino : private serialReceiver::SerialReceiver
     {
       public:
-        CRSF();
-        virtual ~CRSF();
-        void begin();
+        CRSFforArduino();
+        CRSFforArduino(uint8_t RxPin, uint8_t TxPin);
+        ~CRSFforArduino();
+        bool begin();
         void end();
-        void setFrameTime(uint32_t baudRate, uint8_t packetCount = 10);
-        bool receiveFrames(uint8_t rxByte);
-        void getRcChannels(uint16_t *rcChannels);
+        void update();
+        uint16_t getChannel(uint8_t channel);
+        uint16_t rcToUs(uint16_t rc);
+        uint16_t readRcChannel(uint8_t channel, bool raw = false);
+
+        void telemetryWriteBattery(float voltage, float current, uint32_t fuel, uint8_t percent);
+        void telemetryWriteGPS(float latitude, float longitude, float altitude, float speed, float groundCourse, uint8_t satellites);
 
       private:
-        bool rcFrameReceived;
-        uint16_t frameCount;
-        uint32_t timePerFrame;
-        crsfProtocol::frame_t rxFrame;
-        crsfProtocol::frame_t rcChannelsFrame;
-        CRC *crc8;
-        uint8_t calculateFrameCRC();
+        SerialReceiver *_serialReceiver;
     };
-} // namespace serialReceiver
+} // namespace sketchLayer
+
+using namespace sketchLayer;
