+++ src/lib/CRSFforArduino/src/Hardware/CompatibilityTable/CompatibilityTable.cpp
@@ -25,7 +25,7 @@
  * 
  */
 
-#include "CompatibilityTable.h"
+#include "CompatibilityTable.hpp"
 
 namespace hal
 {
