+++ src/lib/CRSFforArduino/src/Hardware/DevBoards/DevBoards.cpp
@@ -24,7 +24,7 @@
  * 
  */
 
-#include "DevBoards.h"
+#include "DevBoards.hpp"
 
 namespace hal
 {
