+++ src/lib/CRSFforArduino/src/SerialReceiver/Telemetry/Telemetry.cpp
@@ -24,7 +24,7 @@
  * 
  */
 
-#include "Telemetry.h"
+#include "Telemetry.hpp"
 
 using namespace crsfProtocol;
